BUGFIX: Make debugger more robust
Prevent exception inception while trying to
render debug outputs.

This can happen when complex objects are encountered
in stacktraces and we try to render a string representation.
Example would be `LazyProps` in Fusion which can easily
throw while evaluating, which is especially nasty as this
can basically recurse if the original error being rendered
occured while trying to render this same LazyProps object.

The suggested fix covers the whole argument rendering in a
try/catch block, as this makes the code even more unreadable
it was refactored to separate methods.