WIP FEATURE: RequestTags
Implements a "RequestTag" system, currently fully exposed on the
ActionRequest but technically we have the data on the HttpRequest for
later changes.

The RequestTags are arbitrary but can be used e.g. as RequestPattern
matcher.

The idea is that you could use the RequestPattern with tags and then
anyone wanting to be part of this AuthenticationProvider can just tag
their route accordingly.
Should also help  with the Frontend vs Backend problem in the
FrontendLogin package.

Other uses TBD, also extensibility (aka other sources for tags).
The implementation in the Router is a bit rough, I wanted to make sure
there is no way to set this from the outside.